Move EncodeXQuery to the xquery package
